requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.1.3
Flask==2.3.3
frozenlist==1.4.0
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.5.0
MarkupSafe==2.1.3
multidict==6.0.4
outcome==1.2.0
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
